Annihilate tombstones early in bookie L-1 cache
We may discard both a tombstone and an active value for some key if
both happen to reside in a L-1 cache. This should eliminate the need
to persist such tombstones to disk only to be discarded much later
during SST merges.

In theory this should improve indexfold performance for such usage
scenarios where there are frequent index changes during key updates.